prøver delete hours

+++ src/main/java/entities/Developer.java
@@ -11,7 +11,7 @@ public class Developer {
     private static final long serialVersionUID = 1L;
     @Id
 
-    private int id;
+    private long id;
     private String name;
     private String email;
     private String phone;
@@ -37,7 +37,7 @@ public class Developer {
     public Developer() {
     }
 
-    public Developer(int id, String name, String email, String phone, String billingPrHour) {
+    public Developer(long id, String name, String email, String phone, String billingPrHour) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -57,11 +57,11 @@ public class Developer {
 //    }
 
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
